requirements.txt
pyobjc-core
pyobjc-framework-CoreSpotlight
pyobjc-framework-CoreServices
pyobjc-framework-UniformTypeIdentifiers
pyinstaller
Pillow
